# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products/{productId}/variant-groups": {
      "get": {
        "operationId": "getPublicProductVariantGroups",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "calculatePriceFromFirstVariantOption": {
                "operationId": "getPublicProductVariantPrice",
                "parameters": {
                  "productId": "$request.path.productId",
                  "optionIds": "$response.body#/0/options/0/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductVariantGroupDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get active variant groups for product",
        "tags": [
          "Product Variants"
        ]
      }
    },
    "/api/products/{productId}/variant-groups/price": {
      "get": {
        "operationId": "getPublicProductVariantPrice",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "optionIds",
            "required": false,
            "in": "query",
            "description": "Selected option ids. Can be repeated (?optionIds=1&optionIds=2) or comma-separated (?optionIds=1,2)",
            "schema": {
              "example": [
                201,
                303
              ],
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantPriceDto"
                }
              }
            }
          }
        },
        "summary": "Calculate final product price by selected option ids",
        "tags": [
          "Product Variants"
        ]
      }
    },
...
  },
  "info": {
    "title": "catalog-service API",
    "description": "REST API + GRPC endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "components": {
    "schemas": {
...
      "ProductVariantOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Variant option id",
            "example": 201
          },
          "name": {
            "type": "string",
            "description": "Variant option title",
            "example": "120 cm"
          },
          "priceDelta": {
            "type": "number",
            "description": "Price delta added to base product price",
            "example": 300
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order",
            "example": 2
          }
        },
        "required": [
          "id",
          "name",
          "priceDelta",
          "sortOrder"
        ]
      },
      "ProductVariantGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Variant group id",
            "example": 12
          },
          "name": {
            "type": "string",
            "description": "Variant group title, for example Size or Material",
            "example": "Size"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Whether one option from this group must be selected",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order",
            "example": 1
          },
          "options": {
            "example": [
              {
                "id": 200,
                "name": "80 cm",
                "priceDelta": 0,
                "sortOrder": 1
              },
              {
                "id": 201,
                "name": "120 cm",
                "priceDelta": 300,
                "sortOrder": 2
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantOptionDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "isRequired",
          "sortOrder",
          "options"
        ]
      },
      "ProductVariantPriceDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "description": "Product id",
            "example": 101
          },
          "basePrice": {
            "type": "number",
            "description": "Base product price",
            "example": 24990
          },
          "optionsPrice": {
            "type": "number",
            "description": "Total delta of selected options",
            "example": 800
          },
          "finalPrice": {
            "type": "number",
            "description": "Final price with selected options",
            "example": 25790
          },
          "selectedOptions": {
            "example": [
              {
                "id": 201,
                "name": "120 cm",
                "priceDelta": 300,
                "sortOrder": 2
              },
              {
                "id": 303,
                "name": "Oak",
                "priceDelta": 500,
                "sortOrder": 1
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantOptionDto"
            }
          }
        },
        "required": [
          "productId",
          "basePrice",
          "optionsPrice",
          "finalPrice",
          "selectedOptions"
        ]
      },
...
    }
  }
}